\name{CCP-package}
\alias{CCP-package}
\alias{CCP}
\docType{package}
\title{Significance Tests for Canonical Correlation Analysis}

\description{
The package provides functions to test for statistical significance
of canonical correlation coefficients, including asymptotic methods
and a Monte Carlo approach. 
Additionally, functions are available to plot the asymptotic distributions
and the permutation distribution, respectively. 
The main user functions are \code{p.asym}, \code{p.perm},
\code{plt.asym}, and \code{plt.perm}.
}

\author{ Uwe Menzel  <uwemenzel@gmail.com>}

\keyword{htest}
\keyword{multivariate}
\keyword{nonparametric}

