\name{CCP-package}
\alias{CCP-package}
\alias{CCP}
\docType{package}
\title{Significance Tests for Canonical Correlation Analysis}

\description{
The package provides functions to test for statistical significance
of canonical correlation coefficients, including asymptotic methods
and a Monte Carlo approach. 
Additionally, functions are available to plot the asymptotic distributions
and the permutation distribution, respectively. 
The main user functions are \code{p.asym}, \code{p.perm},
\code{plt.asym}, and \code{plt.perm}.
}

\details{
\tabular{ll}{
Package: \tab CCP\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2009-12-14\cr
License: \tab GPL\cr
}

}
\author{Uwe Menzel

  Maintainer: Uwe Menzel <uwe.menzel@genpat.uu.se>
}
\keyword{htest}
\keyword{multivariate}
\keyword{nonparametric}

